chore(data): address remaining pompon0 nits on BlockDB replay
Add TODOs for pushing view checks into BlockDB and removing snapshotted
persist cursors, drop inaccurate prune "disk I/O" comments, split recovery
tests into their own file, and run temporary state.Run under scope.Run.
